Add macros to make life easier for the user
Add the "curse" macro, which has 1 variable for the curse name, and must be duplicated each time.
Add the "dynamicEffectsCurse" macro, which allows the user to use this module with Dynamic Effects, and by making it just an active effect of the item, the user only needs the one macro.