Amend control Hook to always clear effects before applying new ones
This is relevant because, on a log out, the effect can be left on an actor. Thus, when logging back in, the actor retains the effect, and if a target is selected first, then there is no controlled token to clear the effects from. If you then control the token, it would add the effect, causing the effect to be doubled